requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
multidict==6.0.4
psycopg2==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
